% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{apply_burnin}
\alias{apply_burnin}
\title{Calculate burnin filter}
\usage{
apply_burnin(table, threshold, burnin)
}
\arguments{
\item{table}{a tibble from `calculate_curve()`}

\item{threshold}{a percentage of the target-source in a sample below which a sample is considered 'discarded'}

\item{burnin}{a number betwen 0 and 1 indicating the fraction of taxa to ignore before applying the threshold}
}
\description{
Calculate burnin filter
}
