% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuperdec-data.R
\docType{data}
\name{cuperdec_database_ex}
\alias{cuperdec_database_ex}
\title{Example isolation source database input for cuperdec}
\format{
An TSV table loaded as a \code{tibble}.
}
\source{
\doi{10.5281/zenodo.3740492}
}
\usage{
data(cuperdec_database_ex)
}
\description{
Example isolation source database used for input to
cuperdec based. Species names are from a NCBI Nt database
and isolation sources gather from the Human Oral Microbiome
database, NCBI GenBank, and manual curation.
}
\examples{

data(cuperdec_database_ex)
load_database(cuperdec_database_ex, target = "oral")
}
\keyword{datasets}
