% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuperdec-data.R
\docType{data}
\name{cuperdec_taxatable_ex}
\alias{cuperdec_taxatable_ex}
\title{Example taxon table input for cuperdec}
\format{
An TSV table loaded as a \code{tibble}.
}
\source{
\doi{10.5281/zenodo.3740492}
}
\usage{
data(cuperdec_taxatable_ex)
}
\description{
Example taxon table used for input to cuperdec based
on data including shotgun-sequenced ancient calculus
samples aligned against the NCBI Nt database from Oct
2017 using MALT. Samples are columns, rows are taxa
and counts are assigned reads.
}
\examples{

data(cuperdec_taxatable_ex)
load_taxa_table(cuperdec_taxatable_ex)
}
\keyword{datasets}
