% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_database}
\alias{load_database}
\title{Load database}
\usage{
load_database(x)
}
\arguments{
\item{x}{a two column TSV file}
}
\description{
This loads a taxon / isolation source database file i.e. first column is
a list of taxa, and the second column is a list of isolation sources.
}
\details{
Taxon names should match that with the taxa table.
}
