% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_taxa_table}
\alias{load_taxa_table}
\title{Load OTU table}
\usage{
load_taxa_table(x)
}
\arguments{
\item{x}{a TSV file or tidy data frame}
}
\description{
This loads a typical taxa table (Samples: columns; Taxa: rows) in TSV format
and standardises some columns, storing the table in the form of a tibble.
}
\examples{
data(cuperdec_taxatable_ex)
taxa_table <- load_taxa_table(cuperdec_taxatable_ex)

}
