% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_taxatable}
\alias{load_taxatable}
\title{Load OTU table}
\usage{
load_taxatable(x)
}
\arguments{
\item{x}{a TSV file}
}
\description{
This loads a typical taxa table (Samples: columns; Taxa: rows) in TSV format
and standardises some columns, storing the table in the form of a tibble.
}
