% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_burnin}
\alias{plot_burnin}
\title{<Internal> Plot simple curves with no sample group separataion but with burnin colouring}
\usage{
plot_burnin(table, burnin_result)
}
\arguments{
\item{table}{output tibble from `calculate_curves()`}

\item{burnin_result}{optional output from `apply_*_burnin()` functions}
}
\description{
Informs `plot_cuperdec()`
}
