% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_grouped}
\alias{plot_grouped}
\title{<Internal> Plot simple curves with sample group separation but no burnin colouring}
\usage{
plot_grouped(table, metadata)
}
\arguments{
\item{table}{output tibble from `calculate_curves()`}

\item{metadata}{optional output from `load_map()`}
}
\description{
Informs `plot_cuperdec()`
}
