% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_grouped_burnin}
\alias{plot_grouped_burnin}
\title{<Internal> Plot simple curves with sample group separation and burnin colouring}
\usage{
plot_grouped_burnin(table, metadata, burnin_result)
}
\arguments{
\item{table}{output tibble from `calculate_curves()`}

\item{metadata}{optional output from `load_map()`}

\item{burnin_result}{optional output from `apply_*_burnin()` functions}
}
\description{
Informs `plot_cuperdec()`
}
