% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_simple}
\alias{plot_simple}
\title{<Internal> Plot simple curves with no metadata or burnin}
\usage{
plot_simple(table)
}
\arguments{
\item{table}{output tibble from `calculate_curves()`}
}
\description{
Informs `plot_cuperdec()`
}
