% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{simple_filter}
\alias{simple_filter}
\title{Apply simple filter}
\usage{
simple_filter(table, threshold)
}
\arguments{
\item{table}{a OTU table loaded with `load_taxa_table()`}

\item{threshold}{a database file loaded with `load_database()`}
}
\description{
Performs the initial decay curve based on percentage of 'target' isolation
source along a rank of most to least abundant taxa for a given sample.
}
