% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{adaptive_burnin_filter}
\alias{adaptive_burnin_filter}
\title{Calculate adaptive burnin retain/discard list}
\usage{
adaptive_burnin_filter(table, threshold)
}
\arguments{
\item{table}{a tibble from `calculate_curve()`}

\item{threshold}{a percentage of the target-source in a sample above which a sample is considered 'retained'}
}
\description{
This function automates a selection of a per-sample 'burn in' based on the
nature of the sample's curve itself (rather than supplying a hard value) by
finding the point from which the 'fluctuation' of the curve doesn't exceed
the mean +- SD of the total curve.
}
