% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_database}
\alias{load_database}
\title{Load database}
\usage{
load_database(x, target)
}
\arguments{
\item{x}{a two column TSV file}

\item{target}{the string in the 'Isolation Source' (i.e. 2nd) column which is the expected target source of the samples}
}
\description{
This loads a taxon / isolation source database file i.e. first column is
a list of taxa, and the second column is a list of isolation sources.
}
\details{
Taxon names should match that with the taxa table.
}
\examples{
data(cuperdec_database_ex)
iso_database <- load_database(cuperdec_database_ex, target = "oral")

}
