% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cuperdec}
\alias{plot_cuperdec}
\title{Plot cumulative percent decay curves}
\usage{
plot_cuperdec(table, metadata)
}
\arguments{
\item{table}{output tibble from `calculate_curves()`}

\item{metadata}{optional output from `load_map()`}
}
\description{
Plot cumulative percent decay curves
}
